undergraduate course page and objects
