+++ CSE_backend/Data_crawling/tag.py
@@ -1,0 +1,10 @@
+# tag create
+
+from api.models import Tag
+
+tag_list = ['수업', '입학', '등록/복학/휴학/재입학', '졸업', '학사 (학부)', '학사 (대학원)', '제2전공/전과', '교환학생/유학',
+'장학', '학부 지도교수', '박사학위예비심사', '자료실', '외부 행사/프로그램', '인턴/취업 (공식 게시)', '기업게시판', '전문연구요원', '미분류']
+
+for i in tag_list :
+    tag = Tag(name = i)
+    tag.save()
