+++ CSE_backend/api/urls.py
@@ -19,7 +19,7 @@ urlpatterns = [
     url(r'^api/tags/$', views.TagList.as_view()),
     url(r'^api/tags/(?P<pk>[0-9]+)/$', views.TagDetail.as_view()),
     url(r'^api/lab/', views.LabList.as_view()),
-
+    url(r'^api/undercourse/', views.UnderCourse.as_view()),
 ]
